requirements.txt
openai==1.101.0
python-dotenv==1.0.0
pelican[markdown]==4.10
Markdown==3.8.2